+++ src/Shell/UpdateShell.php
@@ -16,18 +16,18 @@
  * along with MeYoutube.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author		Mirko Pagliai <mirko.pagliai@gmail.com>
- * @copyright	Copyright (c) 2015, Mirko Pagliai for Nova Atlantis Ltd
+ * @copyright	Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
  * @license		http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link		http://git.novatlantis.it Nova Atlantis Ltd
  */
 namespace MeYoutube\Shell;
 
-use MeTools\Shell\Base\BaseShell;
+use MeTools\Console\Shell;
 
 /**
  * Applies updates
  */
-class UpdateShell extends BaseShell {
+class UpdateShell extends Shell {
 	/**
 	 * Rewrites the header for the shell
 	 */