+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 2.x branch
 ## 2.1 branch
+### 2.1.4
+* added "random videos" widget;
+* added the Install shell.
+
 ### 2.1.3
 * added backward compatibility for old URLs.